requirements.txt
astroid==2.12.12
click==8.1.3
dill==0.3.6
Flask==2.2.2
Flask-Cors==3.0.10
importlib-metadata==5.0.0
isort==5.10.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
lazy-object-proxy==1.8.0
MarkupSafe==2.1.1
mccabe==0.7.0
parso==0.8.3
platformdirs==2.5.4
psycopg==3.1.4
PyJWT==2.6.0
pylint==2.15.5
six==1.16.0
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.4.0
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.10.0
